// src/app/Models/Course.ts
export interface Course {
    courseId: number;
    courseName: string;
    courseThumbnail: string;
    shortDes: string;
    fullDes: string;
    timePublished: Date;
    courseDuration: string;
    oldPrice: number;
    percentSale: number;
    idLevel: number;
    idDetailCate: number;
    profileId:number;
    status:number;
  }